Refactor BuilderEnhancer to only contain helper methods
